tweaks for gcc. swaped index size_t types for uint32_t

+++ cpp/opendnp3/src/opendnp3/IEventBuffer.h
@@ -34,11 +34,11 @@ public:
 
 	virtual ~IEventBuffer() {}
 
-	virtual void Update(const Binary& arEvent, PointClass aClass, size_t aIndex) = 0;
+	virtual void Update(const Binary& arEvent, PointClass aClass, uint32_t aIndex) = 0;
 
-	virtual void Update(const Analog& arEvent, PointClass aClass, size_t aIndex) = 0;
+	virtual void Update(const Analog& arEvent, PointClass aClass, uint32_t aIndex) = 0;
 
-	virtual void Update(const Counter& arEvent, PointClass aClass, size_t aIndex) = 0;
+	virtual void Update(const Counter& arEvent, PointClass aClass, uint32_t aIndex) = 0;
 
 };
